home_inventory/requirements-dev.txt
pytest-django
pytest-cov
mock-django
black
pre-commit
